+++ src/layouts/layout_project_page/layout_project_definition_tab.py
@@ -104,7 +104,6 @@ layout_radio_helper_map_switch = dbc.RadioItems(
         {"label": ProgramDefinitionMapType.SIMPLE.value, "value": ProgramDefinitionMapType.SIMPLE.name},
         {"label": ProgramDefinitionMapType.PROJECTS.value, "value": ProgramDefinitionMapType.PROJECTS.name},
         {"label": ProgramDefinitionMapType.ASSESSMENT_PROBABILITIES.value, "value": ProgramDefinitionMapType.ASSESSMENT_PROBABILITIES.name},
-        {}
     ],
     value="SIMPLE",
     inline=True